Added context of use to narrative of 2 IHI-related extensions
The 2 IHI-related extensions (ihi-record-status and ihi-status) have a
context of use of 'Identifier' set in their respective profiles. This content was
added to the intro narrative markdown pages for the 2 extensions.

The content builds as expected with the IG publisher and there are no
new QA report errors or warnings.