Edited the examples as per the review comments
in the multiple-information-recipents example: as the order of the structured elements is significant, ie extensions appear up the top after id, meta and text, I suggest moving the 4 extensions in the narrative to replicate this ordering
the narrative for the type element has identical content but one has yellow shading and the other not. Suggest making them align, ie add yellow shading to the multiple recipient example
the dates have a different format - also should align; to what exactly?
the author elements have different values - one with just the name and another with all author details. Also should align; maybe keep it simple and have just the name
and maybe the heading shouldn't be "generated narrative" as it wasn't generated. Maybe just "Narrative Details"

CIFMM-1597