Added content in intro.md file
This profile defines a body height information structure for use in an Australian context that tightens the BodyHeightCode slice, aligning to the proposed R4 Vital Signs Body Height profile.

As per the R4 Vital Signs guidance, a separate profile for Body Length is not supported. This observation may be supplemented with an additional coding element containing the LOINC code 8306-3 -Body height - lying (i.e., body length - typically used for infants).