au-practitioner - reverted type references
The constraint of typing the following elements in the Practitioner profile to AU Base profiles has been removed:
- Practitioner.qualification.issuer
- Practitioner.qualification:ahpraRegistration.issuer

The result of which is that they all reference the STU3 counterparts instead. This was the agreement at the HL7 AU working group meetings held 13/9/2018.

No new errors/warnings on QA report.