au-patient - reverted type references
The constraint of typing the following elements in the Patient profile to AU Base profiles has been removed:
- Patient.contact.organization
- Patient.generalPractitioner (as Practitioner and Organization)
- Patient.managingOrganization
- Patient.link.other (as Patient and RelatedPerson)

The result of which is that they all reference the STU3 counterparts instead. This was the agreement at the HL7 AU working group meetings held 13/9/2018.

No new errors/warnings on QA report.