au-healthcareservice - reverted type references
The constraint of typing the following elements in the HealthcareService profile to AU Base profiles has been removed:
- HealthcareService.providedBy
- HealthcareService.location
- HealthcareService.coverageArea

The result of which is that they all reference the STU3 counterparts instead. This was the agreement at the HL7 AU working group meetings held 13/9/2018.

No new errors/warnings on QA report.