au-relatedperson - reverted type references #163
The constraint of typing the following element in the RelatedPerson profile to a AU Base profile has been removed:
- RelatedPerson.patient

The result of which is that it references its STU3 counterpart instead. This was the agreement at the HL7 AU working group meetings held 13/9/2018.

No new errors/warnings on QA report.