au-allergyintolerance - reverted type references
The constraint of typing the following elements in the AllergyIntolerance profile to AU Base profiles has been removed:

- AllergyIntolerance.patient
- AllergyIntolerance.recorder (as patient, practitioner or relatedperson)
- AllergyIntolerance.asserter (as patient, practitioner or relatedperson)

The result of which is that they all reference the STU3 counterparts instead. This was the agreement at the HL7 AU working group meetings held 13/9/2018.

No new errors/warnings on QA report.