Deleted redundant short description
The element AllergyIntolleranbce.reaction had a short value of "Reaction details".  This was previously added as a work-around for a IG Publisher problem where the IG Publisher failed to output a node when there were no changes present.  The IG Publisher issue has since been fixed and the short can be deleted leaving the core description of "Adverse Reaction Events linked to exposure to substance" which is fine.
CIFMM-3596