Minor updates to AU AllergyIntolerance profile (#568)
* Updated binding to updated value set 

The value set "Indicator of Hypersensitivity or Intolerance to Substance" was  updated with a new value set.  The new link was bound to AllergyIntollerance.code and the comment in the intro.md referring to the "soon to be updated valu set" was deleted.
CIFMM-3596.

* Deleted redundant short description

The element AllergyIntolleranbce.reaction had a short value of "Reaction details".  This was previously added as a work-around for a IG Publisher problem where the IG Publisher failed to output a node when there were no changes present.  The IG Publisher issue has since been fixed and the short can be deleted leaving the core description of "Adverse Reaction Events linked to exposure to substance" which is fine.
CIFMM-3596