As per the team review the following changes were agreed to be inserted as a pull request for au-relatedperson
Adding identifier and display for RelatedPerson.patient element of "RelatedPerson-example3" example to show the Implementers that not only reference can be used to mention the patient but also identifier and display can be used
Editing the definition RelatedPerson.communication.language to have "A language which can be used to communicate with about the patient's health." it will look like below :
A language which can be used to communicate with about the patient's health. The ISO-639-1 alpha 2 code in lower case for the language, optionally followed by a hyphen and the ISO-3166-1 alpha 2 code for the region in upper case; e.g. "en" for English, or "en-US" for American English versus "en-EN" for England English.

CIFMM-3758