FBC example removal (#653)
* Removal of poor FBC example

The FBC example had many problems and is better to be removed and re-done at a later date than leave a poor example that would confuse and potentially mislead readers on what a good example should look like.

* Removal of poor FBC example

The FBC example had many problems and is better to be removed and re-done at a later date than leave a poor example that would confuse and potentially mislead readers on what a good example should look like.  Fixes #652.