Ballot fix: intro content - profile status updates and FMM level clean up (#690)
* Changed "FMM Level 0" to "DRAFT 0" and "FMM Level x" to "FMM x"

To align with the content in the guidance "FMM Level 0" was changed to "DRAFT 0" and "FMM Level x" was changed to "FMM x".  This was done for profiles and extensions.
Also, for DRAFT 0 profiles and extensions the watermark was changed/added to be "UNTESTED DRAFT".

* Aligning the merged changes from master with the "DRAFT 0" changes

The merged change from master bought in with it the new profile "AU Base Procedure" which caused a conflict with this branch.  To resolve the conflict these  changes ("FMM Level 0" changed to "DRAFT 0") to align with the content in the guidance.

* Change of StructureDefinition status from Draft to Active

With the extensions both having an FMM 3  and a status of "Draft" which needed to be update to have a status of "Active".

* Reinstated orginals of all old STU3 artefacts 

14 old STU3 extensions inadvertently ported into this R4 IG should instead be removed, rather than have this change applied. Therefore, reinstating these files to their original state.

The removal of these files will be addressed as part of the following GitHub ticket: https://github.com/hl7au/au-fhir-base/issues/667#issuecomment-1022714150

* Extension: Subsidised Concurrent Supply - reduced FMM level to 2

* Subsidised Concurrent Supply extension - update date & patch version

* Indigeneous Status extension - update date & patch version

* Revert more old STU3 intro file to original state

Co-authored-by: David McKillop <david.mckillop@digitalhealth.gov.au>
Co-authored-by: Oridashi <brett.esler@oridashi.com.au>