Change of version to increment from what's in master
The current version of au-smoking status in master is 1.0.1 and this ticket will be the next ticket to be merged into master i.e. it will be  merged before the collector branch is merged.  The collector branch has version of 1.0.2 for au-smokingstatus, but this will need to be incremented when https://stash.digitalhealth.gov.au/projects/CIL/repos/au-fhir-base-r4/pull-requests/216/overview is merged into master.