Deleting List examples that were swapped from GUIDS to non-GUIDS
The identifiers of 3 list examples had the identifiers changed from GUIDs to non-GUID's.  This should have just been a modification of the existing files, but due to the process of swapping out the change from the collector branch the change was reflected as a new file and hence the eixsting files need to be deleted.