Example QA fixes (#722)
* Updates to examples

Five examples were updated including changes to id in the list examples and the change from uuid to local id for the contained resources within practitioner-example4 and practitionerrole-example1.

* Correction to example 

The example have an instance of the Medicare identifier having an incorrect extension associated with it.  The extension has been moved from the Medicare identifier.

* Deleting List examples that were swapped from GUIDS to non-GUIDS

The identifiers of 3 list examples had the identifiers changed from GUIDs to non-GUID's.  This should have just been a modification of the existing files, but due to the process of swapping out the change from the collector branch the change was reflected as a new file and hence the eixsting files need to be deleted.

Co-authored-by: David McKillop <david.mckillop@digitalhealth.gov.au>