Encounter.serviceType and Encounter.type binding update to NCTS content
The binding for Encounter.type was changed to https://healthterminologies.gov.au/fhir/ValueSet/encounter-type-1 with preferred binding.
The binding for Encounter.serviceType was changed to  https://healthterminologies.gov.au/fhir/ValueSet/service-type-1 with preferred binding.  

The following testing was done:
- The valuesets (expanded form) were downloaded from NCTS and created as valuesets (valueset-encounter-type-1.xml & valueset-service-type-1.xml), added to ig.xml and ig.json.
For testing purposes the binding was changed from "preferred" to "required" and the existing encounter examples were modified:
- to include an element from the valueset => PASS
- to not include an element from the valueset => FAIL
Testing outcome was as expected.