Change to au-pathologyresult.category
The structure of the slice for au-pathologyresult.category was changed to reflect the intent of the slice where once category = "laboratory" is mandatory and the other optional category from https://healthterminologies.gov.au/fhir/ValueSet/pathology-diagnostic-service-category-1 with binding = extensible.
Testing done on slice:
1) correct categories: category = "mandatory" and "SR".  expected outcome = pass.  Actual outcome = PASS.
2) No category.  expected outcome = fail. Actual outcome = FAIL.
3) wrong category : category = "imaging" & "SR" (audiology); expected = fail x1 (on imaging); Outcome = FAIL
4) wrong category (slice changed from "extensible" to "required" for testing purposes): category = "imaging" & "AU" (audiology); expected = fail; Outcome = FAIL (found 0)

Minor change to au-imagingresult to correct the order of the elements.