Re-balancing of structureDefinition.description with intro.md content
A request was made to keep the structureDefinition.description brief.  It was found that many were several sentences and the clarifying content beyond the initial definition should be in the intro.md of the respective file.  Hence this commit moves the detailed descriptive content from the structureDefinition.description to the respective intro.md content.