Change of #### to ### in intro.md's and reversal of earlier change
1)  The template output generates a numbering system where the number of #'s for a header determine the number of decimal places/figures associated with the header.
Having 4#'s gives a 5 digit associated header number, having 3#'s gives a 4 digist header number.  When reviewing the output with Brett it was decided to go with 3#'s rather than 4.
Note: the old non-template generation for the IG didn't have the associated header number and hence it wasn't an issue.
Note: the output of the numbering and the header now matches the US Core with a 3 digit profile number and a 4 digit sub heading.
2)  The other change was the reversion of the name> element in the structuredefinition-ihi-verified-date.xml from "IHI Verified Date" to "IHI Verification Date" as Brett seemed to remember that there was a reason for this recent change to "IHI Verification Date".