Template changes updates (#752)
* Removal of <jurisdiction> fragment

The <jurisdiction> fragment has been removed from all resource files.

* Update to intro.md content

In all the intro.md content, the following changes were made:
1) Removal of title with FMM level;
2)  Removal of introdcitory sentence that was already present in the StructureDefinition.description;
3)  Removal of "Extensions" and "Examples".
The following were retained:
a) intro text that wasn't included in the StructureDefinition.description;
b) "Usage Notes" and "Notes on extensions.
Note: Some structureDefinition.descriptions were updated to align with the intro.md content.

* Adding links to data types and resources 

Links were added to data types and resources referenced in structuredefinition.description content in the resources folder.
Plus one intro.md file was updated which was missed in the earlier commit.

* Changes to files in the "intro-notes" folder

All files in the "au-fhir-base-r4/input/intro-notes" folder were prefixed with "StructureDefinition-" so they would be processed by the template build.

* Removal of conflict files

Some files were auto-resolved in the last merge to master and these are now being removed.

* Update to ig.xml

Update to ig.xml including changes to <name> such that name =  "<resource> + <brief description>; update and addition of description where necessary;  adding and updating <exampleCanonical> where necessary; removing "<exampleBoolean value="true"/>

* Typo's, missed content and correction

A number of existing typo's, missing content and miscellaneous corrections have been done

* Typo missed from last commit

Another minor typo in comments that was missed in the last commit.

* Removal of redundant intro.md pages

A number of intro.md pages were deleted as there wasn't a corresponding profile in AU Base.  One had already been moved to AU Core and hence needed to be deleted from AU Base.  
The <meta> was removed from device-example0 as there doesn't exist an au-device profile.

* Re-balancing of structureDefinition.description with intro.md content

A request was made to keep the structureDefinition.description brief.  It was found that many were several sentences and the clarifying content beyond the initial definition should be in the intro.md of the respective file.  Hence this commit moves the detailed descriptive content from the structureDefinition.description to the respective intro.md content.

* Missed changes from previous commit

Changes to these 2 files are a hangover from the previous commit of re-balancing structureDefinition.description with content in intro.md

* Adding url links to structureDefinition.description content 

This commit:
1) adds url links to the structureDefinition.description elements and moves some content into the intro.md;
2)  Updates some intro.mds as some old content was present;
3) removed "example from an ig.xml description to align with all the other examples.

* Aligning name and title with other identifier profiles

The delivery point Identifier profile didn't have "AU" at the beginning of the <name> or <title> elements.  This has been added to align with the other identifier profiles.

* Change of #### to ### in intro.md's and reversal of earlier change

1)  The template output generates a numbering system where the number of #'s for a header determine the number of decimal places/figures associated with the header.
Having 4#'s gives a 5 digit associated header number, having 3#'s gives a 4 digist header number.  When reviewing the output with Brett it was decided to go with 3#'s rather than 4.
Note: the old non-template generation for the IG didn't have the associated header number and hence it wasn't an issue.
Note: the output of the numbering and the header now matches the US Core with a 3 digit profile number and a 4 digit sub heading.
2)  The other change was the reversion of the name> element in the structuredefinition-ihi-verified-date.xml from "IHI Verified Date" to "IHI Verification Date" as Brett seemed to remember that there was a reason for this recent change to "IHI Verification Date".

* Adding missed links

Some links were missed in the StructureDefinition-au-healthcarecardnumber-intro.md.

* Removal of blank lines from ig.xml

A previous bulk edit of the ig.xml where an element (? <GroupId>) was removed, a blank line remained.  These blank lines have been deleted.

* Corrections following peer review

- The Condition.category was lost since Brett added it on the 22/9/2022 - it is now added back in.
- deleted "parity-example0 as it should be in AU Core
-  other typo peer review corrections.

* Changes to ig.xml following peer review

The agreed format for the <<name> element in the ig.xml examples was  "<resource> - <short description>", hence 4 instances of "Organisation" were changed to "Organization".

* Corrections following peer review

Some corrections to the ig.xml  including corrections to descriptions and a correction to a typo>
Plus corrections to incorrect links and removal of a duplicate term.